+++ src/display.c
@@ -58,14 +58,16 @@ void display_serialize(FILE *rsp, uint32_t did)
             "\t\"index\":%d,\n"
             "\t\"label\":\"%s\",\n"
             "\t\"frame\":{\n\t\t\"x\":%.4f,\n\t\t\"y\":%.4f,\n\t\t\"w\":%.4f,\n\t\t\"h\":%.4f\n\t},\n"
-            "\t\"spaces\":[%s]\n"
+            "\t\"spaces\":[%s],\n"
+            "\t\"has-focus\":%s\n"
             "}",
             did,
             uuid ? uuid : "<unknown>",
             display_manager_display_id_arrangement(did),
             display_label ? display_label->label : "",
             frame.origin.x, frame.origin.y, frame.size.width, frame.size.height,
-            buffer);
+            buffer,
+            json_bool(did == g_display_manager.current_display_id));
 }
 
 CFStringRef display_uuid(uint32_t did)